Refactor hidden field clearing logic into a separate utility function
